requirements.txt
Django==4.1.2
djangorestframework==3.14.0
qrcode==7.3.1
gunicorn==20.1.0
Pillow==9.2.0